Start categorizing model items
