+++ lib/model/pattern/note.dart
