+++ lib/widgets/editors/arranger/pattern_picker/pattern_picker_cubit.dart
@@ -20,14 +20,12 @@
 import 'dart:async';
 
 import 'package:anthem/commands/pattern_commands.dart';
-import 'package:anthem/main.dart';
+import 'package:anthem/commands/state_changes.dart';
+import 'package:anthem/model/pattern/pattern.dart';
+import 'package:anthem/model/project.dart';
+import 'package:anthem/model/store.dart';
 import 'package:bloc/bloc.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
-
-import '../../../../commands/state_changes.dart';
-import '../../../../model/pattern.dart';
-import '../../../../model/project.dart';
-import '../../../../model/store.dart';
 
 part 'pattern_picker_state.dart';
 part 'pattern_picker_cubit.freezed.dart';
