+++ lib/widgets/editors/piano_roll/piano_roll_cubit.dart
@@ -23,13 +23,12 @@ import 'dart:math';
 import 'package:anthem/commands/pattern_commands.dart';
 import 'package:anthem/commands/state_changes.dart';
 import 'package:anthem/helpers/get_id.dart';
-import 'package:anthem/model/note.dart';
+import 'package:anthem/model/pattern/note.dart';
 import 'package:anthem/model/project.dart';
+import 'package:anthem/model/shared/time_signature.dart';
 import 'package:anthem/model/store.dart';
 import 'package:bloc/bloc.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
-
-import '../../../model/time_signature.dart';
 
 part 'piano_roll_state.dart';
 part 'piano_roll_cubit.freezed.dart';
