+++ lib/widgets/editors/shared/timeline_cubit.dart
@@ -18,13 +18,12 @@
 */
 
 import 'package:anthem/commands/state_changes.dart';
+import 'package:anthem/model/pattern/pattern.dart';
+import 'package:anthem/model/project.dart';
+import 'package:anthem/model/shared/time_signature.dart';
+import 'package:anthem/model/store.dart';
 import 'package:bloc/bloc.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
-
-import '../../../model/pattern.dart';
-import '../../../model/project.dart';
-import '../../../model/store.dart';
-import '../../../model/time_signature.dart';
 
 part 'timeline_state.dart';
 part 'timeline_cubit.freezed.dart';
